Shorten introductory material.
Removed most of the background material so as to not confuse the elements
that fall within the scope of the milestone and that which does not.
Instead, discuss only what is within the confines of the milestone so that
extranious analysis that we do not cover cannot be used as failure.