Improve figure placement and a few other changes.
The automatic figure placement is working pretty poorly for this document.
There are many large figures, so that they tend to be placed later in the
document than where they are referenced.  To fix that, move some definitions
ahead and force figures to sometimes flush before continuing text.  We
should keep on top of this for future changes in the document.